processor-api/requirements.txt
flask
boto3
prometheus_client
PyPDF2
flask_injector
huggingface_hub
python-dotenv
pika